// test/utils/prisma.types.d.ts
import { Prisma } from '@prisma/client';

export type PrismaTypes = {
  asso: Prisma.AssoCreateInput,
  assoKeyword: Prisma.AssoKeywordCreateInput,
  assoMembership: Prisma.AssoMembershipCreateInput,
  assoMembershipPermission: Prisma.AssoMembershipPermissionCreateInput,
  assoMembershipRole: Prisma.AssoMembershipRoleCreateInput,
  assoMessage: Prisma.AssoMessageCreateInput,
  event: Prisma.EventCreateInput,
  eventAnswer: Prisma.EventAnswerCreateInput,
  eventCategory: Prisma.EventCategoryCreateInput,
  eventPrivacy: Prisma.EventPrivacyCreateInput,
  gitHubIssue: Prisma.GitHubIssueCreateInput,
  semester: Prisma.SemesterCreateInput,
  timetableGroup: Prisma.TimetableGroupCreateInput,
  timetableEntry: Prisma.TimetableEntryCreateInput,
  timetableEntryOverride: Prisma.TimetableEntryOverrideCreateInput,
  translation: Prisma.TranslationCreateInput,
  uE: Prisma.UECreateInput,
  uEAnnal: Prisma.UEAnnalCreateInput,
  uEAnnalReport: Prisma.UEAnnalReportCreateInput,
  uEAnnalReportReason: Prisma.UEAnnalReportReasonCreateInput,
  uEComment: Prisma.UECommentCreateInput,
  uECommentReply: Prisma.UECommentReplyCreateInput,
  uECommentReport: Prisma.UECommentReportCreateInput,
  uECommentReportReason: Prisma.UECommentReportReasonCreateInput,
  uECourse: Prisma.UECourseCreateInput,
  uECourseExchange: Prisma.UECourseExchangeCreateInput,
  uECourseExchangeReply: Prisma.UECourseExchangeReplyCreateInput,
  uECredit: Prisma.UECreditCreateInput,
  uECreditCategory: Prisma.UECreditCategoryCreateInput,
  uEInfo: Prisma.UEInfoCreateInput,
  uEStarCriterion: Prisma.UEStarCriterionCreateInput,
  uEStarVote: Prisma.UEStarVoteCreateInput,
  uEWorkTime: Prisma.UEWorkTimeCreateInput,
  user: Prisma.UserCreateInput,
  userAddress: Prisma.UserAddressCreateInput,
  userBan: Prisma.UserBanCreateInput,
  userBDEContribution: Prisma.UserBDEContributionCreateInput,
  userBranche: Prisma.UserBrancheCreateInput,
  userEtuUTTTeam: Prisma.UserEtuUTTTeamCreateInput,
  userFormation: Prisma.UserFormationCreateInput,
  userInfos: Prisma.UserInfosCreateInput,
  userMailsPhones: Prisma.UserMailsPhonesCreateInput,
  userOtherAttribut: Prisma.UserOtherAttributCreateInput,
  userOtherAttributValue: Prisma.UserOtherAttributValueCreateInput,
  userPreference: Prisma.UserPreferenceCreateInput,
  userRGPD: Prisma.UserRGPDCreateInput,
  userSocialNetwork: Prisma.UserSocialNetworkCreateInput,
  userTimestamps: Prisma.UserTimestampsCreateInput,
  userTimetableGroup: Prisma.UserTimetableGroupCreateInput,
  userUESubscription: Prisma.UserUESubscriptionCreateInput,
  uTTBranche: Prisma.UTTBrancheCreateInput,
  uTTFiliere: Prisma.UTTFiliereCreateInput,
  uTTFormation: Prisma.UTTFormationCreateInput,
  uTTFormationFollowingMethod: Prisma.UTTFormationFollowingMethodCreateInput,
};